Add custom configs for iTerm on macOS
Allows for default configs and local overrides to control iTerm-related
settings. This stems from an issue where external applications hooking
into iTerm to start a session and run a command were not working; when
the ~/.iterm2/it2setkeylabel command is called it would cause the
calling command to not end up being run when the shell was finished
loading (this was first noticed when trying to open a connection in
iTerm from XPipe).

To fix this in cases where the touchbar functionality is not needed,
while still keeping the functionality available, the config var
DOTFILES_CONFIG_MACOS_ITERM_TOUCHBAR has been added, which can be
overridden in the _inc/config_macOS.custom file.

Additionally, a DOTFILES_CONFIG_MACOS_ITERM_SHELL_INTEGRATION config var
has been added to control loading the iTerm2 shell integration script.
No issues have been noticed with this script so far, but the config
allows for some more personal control as desired.